+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Installation Guide: a more detailed guide, that better explains the installation of pot
 - fdescfs attribute: new attribute to mount fdescfs inside the pot
+- init: add a network configuration validation step
 
 ### Changed
 - POT_CACHE: the cache used by import is now a dataset, child of POT_ZFS_ROOT